4.15. -4.23.

+++ part4/blogilista/controllers/blogsRouter.js
@@ -1,19 +1,58 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-    response.json(blogs)
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const savedBlog = await new Blog(request.body).save()
+  
+  const user = request.user
+  if (!user) {
+    return response.status(400).json({ error: 'UserId missing or not valid' })
+  }
+
+  const { title, author, url, likes } = request.body
+
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+  
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'UserId missing or not valid'})
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found'})
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'unauthorized: not the creator of this blog' })
+  }
+
+  await Promise.all([
+    Blog.findByIdAndDelete(request.params.id),
+    User.findByIdAndUpdate(user._id, { $pull: { blogs: blog._id } })
+  ])
   response.status(204).end()
 })
